+++ falcon2/src/old/index.ts
@@ -3,4 +3,3 @@ export * from "./basic";
 export * from "./config";
 export * from "./db/index";
 export * from "./util";
-export * from "./core/index";
